chore: refactor fold's structure
