+++ dubbo-samples-migration/dubbo-samples-migration-api/src/main/java/org/apache/dubbo/migration/api/GreeterService.java
@@ -14,10 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.migration;
-
-import org.apache.dubbo.migration.pojo.HelloReply;
-import org.apache.dubbo.migration.pojo.HelloRequest;
+package org.apache.dubbo.migration.api;
 
 public interface GreeterService {
