+++ dubbo-samples-migration/dubbo-samples-migration-api/src/main/java/org/apache/dubbo/migration/api/HelloRequest.java
@@ -1,4 +1,4 @@
-package org.apache.dubbo.migration.pojo;
+package org.apache.dubbo.migration.api;
 
 import java.io.Serializable;
