+++ dubbo-samples-migration/pom.xml
@@ -10,14 +10,15 @@
     <version>1.0-SNAPSHOT</version>
     <modules>
         <module>dubbo-samples-migration-default</module>
+        <module>dubbo-samples-migration-api</module>
     </modules>
     
     <properties>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
-<!--        <dubbo.version>3.0.2.1</dubbo.version>-->
+        <dubbo.version>3.0.2.1</dubbo.version>
 <!--        <dubbo.version>3.0.3-SNAPSHOT</dubbo.version>-->
-        <dubbo.version>2.7.13</dubbo.version>
+<!--        <dubbo.version>2.7.13</dubbo.version>-->
         <junit.version>4.12</junit.version>
         <maven-compiler-plugin.version>3.7.0</maven-compiler-plugin.version>
         <source.level>1.8</source.level>
